fix: resolve circular import
