+++ pybedlite/bed_record.py
@@ -18,9 +18,6 @@ from typing import Optional
 from typing import Tuple
 from typing import List
 from typing import ClassVar
-from typing import Type
-
-from pybedlite.overlap_detector import Interval
 
 
 """Maximum BED fields that can be present in a well formed BED file written to specification"""
@@ -192,45 +189,3 @@ class BedRecord:
         )
         fields = self.bed_fields[:number_of_output_fields]
         return "\t".join(fields)
-
-    def to_interval(self) -> Interval:
-        """
-        Convert a `BedRecord` to an `Interval` instance.
-
-        Note that when the `BedRecord` does not have a specified strand, the `Interval`'s negative
-        attribute is set to False. This mimics the behavior of `OverlapDetector.from_bed()` when
-        reading a record that does not have a specified strand.
-
-        Returns:
-            An `Interval` corresponding to the same region specified in the record.
-        """
-        return Interval(
-            refname=self.chrom,
-            start=self.start,
-            end=self.end,
-            negative=self.strand is BedStrand.Negative,
-            name=self.name,
-        )
-
-    @classmethod
-    def from_interval(cls: Type["BedRecord"], interval: Interval) -> "BedRecord":
-        """
-        Construct a `BedRecord` instance from an `Interval` instance.
-
-        **Note that `Interval` cannot represent a `BedRecord` with a missing strand.**
-        Converting a record with no strand to `Interval` and then back to `BedRecord` will result in
-        a record with **positive strand**.
-
-        Args:
-            interval: The `Interval` instance to convert.
-
-        Returns:
-            A `BedRecord` corresponding to the same region specified in the interval.
-        """
-        return cls(
-            chrom=interval.refname,
-            start=interval.start,
-            end=interval.end,
-            strand=BedStrand.Negative if interval.negative else BedStrand.Positive,
-            name=interval.name,
-        )
